Add client support for player list
